Basic Creature Needs UI Working
Can now start the UI counting down when the player clocks in and have the display update on the wall.

Need to get this started from Main when I'm initialising the days instead. Would be great to be able to debug each day with the press of a button.

Consider storing this information in resources