Control Nodes Refactored into Station Nodes
Stations Nodes now contain unhandled input callbacks that can be activated and deactivated by the Stations header when the player interacts with a given Station.

Potentially bug with the inputs registering the release of the interact key, but will see if that has an effect before acting. Could either delay input activation until animation is finished, or create a Timer that will not permit input until it has timed out