Cleaning Station Running
Cleaning Station is now basically complete. It uses the same logic as the Feeding Station. Will adjust the appearance of the buttons and the request images to give it its own unique vibe though.

TO DO: Got a bug with the new request system, where it will only make a new request if the current value is UNDER the threshold, which is the opposite of what I want. Saving before I attempt a debug