Standardised Levels and Intros for Future
Each level now opens with a black screen and potentially a harsh cut to a title card, though I'm thinking in the creature room this doesn't happen until the player walks through the main door.

Going to create the sequence where player enters lobby, picks up card, gets on elevator. (Need to be able to try exit door as well)